web/requirements.txt
Flask==3.0.3
Flask-RESTful==0.3.9
Flask-Cors==3.0.10
requests==2.31.0
h5py==3.10.0
google-cloud-storage==2.17.0
gunicorn==22.0.0
python-dotenv==1.0.1
scquill==0.2.4
mygene==3.2.2